bookrecog_server/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
abseil-cpp=20200225.2=he1b5a44_2
absl-py=0.11.0=py38h578d9bd_0
# addict=2.4.0=pypi_0
aiohttp=3.7.3=py38h497a2fe_0
albumentations=0.5.2=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
attrs=20.3.0=pyhd3deb0d_0
blas=2.21=mkl
blinker=1.4=py_1
blosc=1.20.1=h9c3ff4c_0
brotli=1.0.9=he1b5a44_3
brotlipy=0.7.0=py38h8df0ef7_1001
brunsli=0.1=h9c3ff4c_0
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cachetools=4.1.1=py_0
cairo=1.16.0=h7979940_1007
certifi=2020.12.5=py38h578d9bd_0
cffi=1.14.4=py38ha65f79e_1
chardet=3.0.4=py38h924ce5b_1008
charls=2.1.0=h9c3ff4c_2
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.3.1=py38h2b97feb_0
cub=1.8.0=owncuda10.2
cudatoolkit=10.2.89=h8f6ccaa_6
cycler=0.10.0=py_2
cytoolz=0.11.0=py38h25fe258_1
dask-core=2020.12.0=pyhd8ed1ab_0
dbus=1.13.6=hfdff14a_1
decorator=4.4.2=py_0
easydict=1.9=py_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h3215721_1
flask=1.1.2=pyh9f0ad1d_0
flask_cors=3.0.9=pyh9f0ad1d_0
fontconfig=2.13.1=h736d332_1003
freetype=2.10.4=h7ca028e_0
geos=3.8.1=he1b5a44_0
gettext=0.19.8.1=h0b5b191_1005
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h36c2ea0_2
glib=2.66.4=hcd2ae1e_1
gmp=6.2.0=own
gnutls=3.6.13=h85f3911_1
google-auth=1.24.0=pyhd3deb0d_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.13=h58526e2_1001
grpc=1.30.0=own
grpcio=1.34.0=py38hdd6454d_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.18.2=ha23517c_0
harfbuzz=2.7.2=h5cf4720_2
hdf5=1.10.6=nompi_h6a2412b_1112
icu=68.1=h58526e2_0
idna=2.10=pyh9f0ad1d_0
imagecodecs=2020.5.30=py38h67f9a0b_4
imageio=2.9.0=py_0
imgaug=0.4.0=py_1
importlib-metadata=3.3.0=py38h578d9bd_2
itsdangerous=1.1.0=py_0
jasper=1.900.1=h07fcdf6_1006
jinja2=2.11.2=pyh9f0ad1d_0
jpeg=9d=h36c2ea0_0
jxrlib=1.1=h7f98852_2
kiwisolver=1.3.1=py38h82cb98a_0
krb5=1.17.2=h926e7f8_0
lame=3.100=h14c3975_1001
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35.1=hea4e1c9_1
lerc=2.2=h9c3ff4c_0
libaec=1.0.4=h9c3ff4c_1
libblas=3.8.0=21_mkl
libcblas=3.8.0=21_mkl
libclang=11.0.0=default_ha5c780c_2
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libglib=2.66.4=h164308a_1
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
liblapack=3.8.0=21_mkl
liblapacke=3.8.0=21_mkl
libllvm11=11.0.0=he513fc3_0
libnghttp2=1.41.0=h8cfc5f6_2
libopencv=4.5.0=py38_4
libpng=1.6.37=h21135ba_2
libpq=12.3=h255efa7_3
libprotobuf=3.14.0=h780b84a_0
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.2.0=hdc55705_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.1.0=h36c2ea0_3
libxcb=1.13=h14c3975_1002
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.10=h72842e0_3
libzopfli=1.0.3=h9c3ff4c_0
llvm-openmp=11.0.0=hfc4b9b4_1
lz4-c=1.9.2=he1b5a44_3
magma=2.5.2=owncuda10.2
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py38h8df0ef7_2
matplotlib-base=3.3.3=py38h5c7f4ab_0
mkl=2020.4=h726a3e6_304
# mmcls=0.6.0=dev_0
# mmcv=1.1.5=pypi_0
multidict=5.1.0=py38h497a2fe_0
mysql-common=8.0.22=ha770c72_1
mysql-libs=8.0.22=h1fd7589_1
nccl=2.8.3=owncuda10.2
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
networkx=2.5=py_0
ninja=1.10.2=h4bd325d_0
nspr=4.29=he1b5a44_1
nss=3.60=hb5efdd6_0
numpy=1.19.4=py38h18fd61f_2
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
openblas=0.3.7=own
opencv=4.5.0=py38_4
openh264=2.1.1=h8b12597_0
openjpeg=2.3.1=hf7af979_3
openssl=1.1.1i=h7f98852_0
pcre=8.44=he1b5a44_0
pillow=8.0.1=py38h70fbd49_0
pip=20.3.3=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
protobuf=3.14.0=py38h709712a_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=4.5.0=py38h81c977d_4
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyha679756_2000
pyjwt=1.7.1=py_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h924ce5b_2
python=3.8.6=hffdb5ce_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
# pytorch=1.6.0=py3.8_cuda10.2.89_cudnn7.6.5_0
pywavelets=1.1.1=py38h5c078b8_3
pyyaml=5.3.1=py38h8df0ef7_1
qt=5.12.9=h9d6b050_2
readline=8.0=he28a2e2_2
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
scikit-image=0.17.2=py38h51da96c_4
scipy=1.5.3=py38hb2138dd_0
setuptools=50.3.2=py38ha679756_2000
shapely=1.7.1=py38hc7361b7_1
shellingham=1.3.2=py_0
six=1.15.0=pyh9f0ad1d_0
snappy=1.1.8=he1b5a44_3
sqlite=3.34.0=h74cdb3f_0
tensorboard=2.4.0=pyhd8ed1ab_0
tensorboard-plugin-wit=1.7.0=pyh9f0ad1d_0
tifffile=2020.12.8=pyhd8ed1ab_0
tk=8.6.10=h21135ba_1
toolz=0.11.1=py_0
torchaudio=0.6.0=py38
torchvision=0.7.0=py38_cu102
tornado=6.1=py38h25fe258_0
tqdm=4.54.1=pyhd8ed1ab_1
typer=0.3.1=py_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd8ed1ab_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3deb0d_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
# yapf=0.30.0=pypi_0
yarl=1.6.3=py38h497a2fe_0
zfp=0.5.5=h9c3ff4c_4
zipp=3.4.0=py_0
zlib=1.2.11=own
zstd=1.4.5=h6597ccf_2
